refactor(gui): extract session-bound wiring from constructor (#3351 Phase 2c)
Ninth PR of the #3351 series, third of the constructor decomposition —
and the strategically central one: the three blocks that constitute
"a connected radio" move into a new MainWindow_Session.cpp, which is
explicitly the seed file for the RadioSession aggregate (#3445 Camp B).

Same pattern as 2a/2b: each block becomes a named private wireXxx()
method called at exactly the original constructor position. Three
additive MainWindow.h declarations.

- wireDiscovery() (~220 lines): LAN discovery, heartbeat / TCP-ping
  disconnect detection, SmartLink (WAN).
- wireRadioModel() (~450 lines): RadioModel connection-state + status
  routing, TX audio stream IDs for the DAX TX path.
- wirePanLifecycle() (~455 lines): panadapter stream → spectrum widgets,
  S-history markers, multi-pan create/destroy/rearrange lifecycle.

The TU's header comment marks it as the RadioSession seed and warns
against adding window-chrome wiring to it.

Supporting moves, called out for review:

- Moved with their only callers into the Session TU anon namespace:
  defaultPanLayoutForCount, logXvtrWaterfallDecision (+ its log-interval
  constant).
- Promoted to MainWindowHelpers.h (now cross-TU):
  kPanLayoutRestoreWaitingForFirstPan / kPanLayoutRestoreWindowMs —
  shared between the connect-time restore logic (MainWindow.cpp) and the
  moved multi-pan lifecycle.

Includes carried explicitly (Qt 6.4.2-floor discipline), including the
platform-conditional DaxBridge headers.

MainWindow.cpp: 9,954 → 8,759 lines (series: 19,474 → 8,759, −55.0%).
Constructor: ~2,800 → ~1,700.
Verified: full build clean, all 33 ctest suites pass, a11y lint clean.

Refs #3351 #3445. Principle XI.
