+++ CMakeLists.txt
@@ -637,6 +637,7 @@ set(GUI_SOURCES
     src/gui/MainWindowHelpers.cpp
     src/gui/MainWindow_Controllers.cpp
     src/gui/MainWindow_Menus.cpp
+    src/gui/MainWindow_Session.cpp
     src/gui/MainWindow_Shortcuts.cpp
     src/gui/MainWindow_DigitalModes.cpp
     src/gui/MainWindow_Spots.cpp
